Split out tests into different files based on entity or span group behavior.
